Fix overflow when evaluating passed pawns
Results of engine1 vs engine2 (8+0.08, NULL, NULL, 8moves_v3.pgn):
Elo: 4.30 +/- 9.47, nElo: 5.24 +/- 11.55
LOS: 81.32 %, DrawRatio: 37.90 %, PairsRatio: 1.02
Games: 3478, Wins: 1326, Losses: 1283, Draws: 869, Points: 1760.5 (50.62 %)
Ptnml(0-2): [194, 340, 659, 321, 225], WL/DD Ratio: 5.34
LLR: 2.94 (-2.94, 2.94) [-10.00, 0.00]

+++ src/evaluate.rs
@@ -251,12 +251,12 @@ impl Board {
             / (MIN_GAME_STAGE_FULLY_MIDGAME);
 
         let white_passed = self.white_passed_pawns();
-        let white_passed_distance = (south_fill(white_passed) &!white_passed).count_ones();
+        let white_passed_distance = (south_fill(white_passed) &!white_passed).count_ones() as i16;
         
         let black_passed = self.black_passed_pawns();
-        let black_passed_distance = (north_fill(black_passed) &!black_passed).count_ones();
+        let black_passed_distance = (north_fill(black_passed) &!black_passed).count_ones() as i16;
 
-        let net_passed_pawns = (white_passed_distance - black_passed_distance) as i16;
+        let net_passed_pawns = white_passed_distance - black_passed_distance;
 
         // Add a small variance to try to avoid repetition
         // isolated_pawns * ISOLATED_PAWN_PENALTY.get()
